examples/adk_a2a_chat/requirements.txt
# Core dependencies
agentic-kernel>=0.2.10
google-adk>=0.1.1
google-generativeai>=0.3.0
pydantic>=2.6.3
python-dotenv>=1.1.0

# Networking and async
aiohttp>=3.11.14
httpx>=0.28.1
starlette>=0.41.3
uvicorn>=0.30.0

# UI
rich>=14.0.0
prompt_toolkit>=3.0.43
mesop>=0.0.1

# Testing
pytest>=8.3.5
pytest-asyncio>=0.26.0
